third_party/blink/web_tests/svg/dynamic-updates/SVGFECompositeElement-svgdom-in-prop-expected.txt

SVG 1.1 dynamic update tests

Tests dynamic updates of the 'in' property of the SVGFECompositeElement object

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS overComposite1.in1.baseVal is "F12"
PASS overComposite1.in1.baseVal is "F1"
PASS successfullyParsed is true

TEST COMPLETE
